# web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mosic 2.0 API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/fascicoli": {
            "get": {
                "tags": [
                    "Fascicoli"
                ],
                "summary": "Lista fascicoli",
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo",
                        "examples": {
                            "application/json": {
                                "response": 200,
                                "total_results": 1326,
                                "filter_results": 1336,
                                "limit": "99999",
                                "offset": 0,
                                "data": [
                                    {
                                        "id": "1325",
                                        "codice_repertorio": "0",
                                        "id_titolari": "2",
                                        "numero_fascicolo": "0",
                                        "argomento": "qwerty",
                                        "id_amministrazione": "0",
                                        "data_magazzino": "2017-01-20",
                                        "data_cipe": "0000-00-00",
                                        "data_cipe2": "0000-00-00",
                                        "archiviazione_repertorio": "0",
                                        "annotazioni": "qwerty",
                                        "id_numeri_delibera": "0",
                                        "id_esiti_cipe": "0",
                                        "id_archivio_repertorio": "0",
                                        "id_amministrazioni": "5,4,1",
                                        "id_tags": [
                                            16,
                                            15,
                                            14
                                        ]
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    }
                }
            },
            "post": {
                "tags": [
                    "Fascicoli"
                ],
                "summary": "Creazione fascicolo",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo"
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    }
                }
            }
        },
...
    },
...
}